refactor(theme): themeInfo 가져오는 로직 useQuery로 변경
